Implement append_text_to_file function
Added the append_text_to_file function in 2-append_text_to_file.c following the provided prototype and Betty coding style. This function appends the NULL-terminated string from the text_content parameter to the end of the file specified by filename.The function returns 1 on success and -1 on failure. If filename is NULL, the function returns -1. If text_content is NULL, nothing is added to the file, and the function returns 1 if the file exists or -1 if the file does not exist or if there are no required permissions to write the file.